Edited Player.cpp (main character can now move)

+++ sfml-game-template-master/projects/Block.h
@@ -33,8 +33,8 @@ class Block{
          * these are used to find the values of a particular block
          */
          Sprite sprite() { return *_sprite; }
-         int xPos() { return _xPos; }
-         int yPos() { return _yPos; }
+         float xPos() { return _xPos; }
+         float yPos() { return _yPos; }
          bool passable() { return _passable; }
          // END ACCESSOR METHODS
         
@@ -51,9 +51,9 @@ class Block{
          Sprite* _sprite; // the sprite itself
 
          bool _passable; // true if a character is able to walk on it
-         int _xPos; // the x position of the block within the grid
-         int _yPos; // the y pos of the block within the grid
-         int _width; // the pixle width and height of the block
+         float _xPos; // the x position of the block within the grid
+         float _yPos; // the y pos of the block within the grid
+         float _width; // the pixle width and height of the block
 
 };
 #endif /* Block_hpp */
